medium_apis/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
flask-cors==4.0.0 